+++ riscv/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Re-use `try_*` functions in `mcountinhibit`
 - Use CSR helper macros to define `mcause` register
 - Use CSR helper macros to define `medeleg` register
+- Use CSR helper macros to define `mideleg` register
 
 ## [v0.12.1] - 2024-10-20